Rossman Sales Prediction in Python
The model build is based on the Linear Regression and Random Forest algorithm. The root mean squared error is used to determine the accuracy of a forecast. The model is also evaluated based on the level of insight provided by the visualizations created, in addition to the RMSE value. The R-Squared values aid in the evaluation of the model created.